ClaudeDesktopBridge/requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
requests==2.31.0
python-dotenv==1.0.0
pywin32==306; sys_platform == 'win32'

# Claude API
anthropic==0.5.0

# Computer control
pyautogui==0.9.54
pillow==10.0.0
keyboard==0.13.5
mouse==0.7.1
pynput==1.7.6
psutil==5.9.5

# Web automation
selenium==4.10.0
webdriver-manager==3.8.6

# GUI
streamlit>=1.29.0

# Testing
pytest==7.3.1
httpx==0.24.1
